requirements.txt
numpy
tqdm
voyageai
anthropic
google-generativeai
groq
python-dotenv
uvicorn
fastapi